+++ packages/@aws-cdk/tmp-toolkit-helpers/package.json
@@ -78,7 +78,8 @@
   "exports": {
     ".": "./lib/index.js",
     "./package.json": "./package.json",
-    "./api": "./lib/api/index.js"
+    "./api": "./lib/api/index.js",
+    "./util": "./lib/util/index.js"
   },
   "//": "~~ Generated by projen. To modify, edit .projenrc.js and run \"npx projen\"."
 }
