+++ packages/@aws-cdk/tmp-toolkit-helpers/src/util/index.ts
@@ -4,7 +4,9 @@ export * from './bool';
 export * from './bytes';
 export * from './cloudformation';
 export * from './content-hash';
+export * from './directories';
 export * from './format-error';
+export * from './json';
 export * from './objects';
 export * from './parallel';
 export * from './serialize';
