+++ packages/@aws-cdk/toolkit-lib/.projen/tasks.json
@@ -181,6 +181,12 @@
         },
         {
           "exec": "node ./lib/api/aws-cdk.js >/dev/null 2>/dev/null </dev/null"
+        },
+        {
+          "exec": "node ./lib/api/shared-public.js >/dev/null 2>/dev/null </dev/null"
+        },
+        {
+          "exec": "node ./lib/private/util.js >/dev/null 2>/dev/null </dev/null"
         }
       ]
     },
