+++ packages/@aws-cdk/toolkit-lib/lib/api/shared-public.ts
@@ -1,4 +1,3 @@
 /* eslint-disable import/no-restricted-paths */
 
-// APIs
 export * from '../../../tmp-toolkit-helpers/src/api';
