+++ packages/aws-cdk/lib/api/aws-auth/user-agent.ts
@@ -1,6 +1,6 @@
 import * as path from 'path';
 import { readIfPossible } from './util';
-import { rootDir } from '../../util/directories';
+import { cliRootDir } from '../../cli/root-dir';
 
 /**
  * Find the package.json from the main toolkit.
@@ -9,7 +9,7 @@ import { rootDir } from '../../util/directories';
  * Fall back to argv[1], or a standard string if that is undefined for some reason.
  */
 export function defaultCliUserAgent() {
-  const root = rootDir(false);
+  const root = cliRootDir(false);
   const pkg = JSON.parse((root ? readIfPossible(path.join(root, 'package.json')) : undefined) ?? '{}');
   const name = pkg.name ?? path.basename(process.argv[1] ?? 'cdk-cli');
   const version = pkg.version ?? '<unknown>';
