+++ packages/aws-cdk/lib/api/deployments/cloudformation.ts
@@ -16,8 +16,7 @@ import type { Deployments } from './deployments';
 import { debug } from '../../cli/messages';
 import { IoMessaging } from '../../toolkit/cli-io-host';
 import { ToolkitError } from '../../toolkit/error';
-import { formatErrorMessage } from '../../util/format-error';
-import { deserializeStructure } from '../../util/serialize';
+import { formatErrorMessage, deserializeStructure } from '../../util';
 import type { ICloudFormationClient, SdkProvider } from '../aws-auth';
 import { StackStatus } from '../stack-events';
 import { makeBodyParameter, TemplateBodyParameter } from '../util/template-body-parameter';
