+++ packages/aws-cdk/lib/api/deployments/nested-stack-helpers.ts
@@ -3,7 +3,7 @@ import type { CloudFormationStackArtifact } from '@aws-cdk/cx-api';
 import * as fs from 'fs-extra';
 import type { SDK } from '../aws-auth';
 import { CloudFormationStack, type Template } from './cloudformation';
-import { formatErrorMessage } from '../../util/format-error';
+import { formatErrorMessage } from '../../util';
 import { LazyListStackResources, type ListStackResources } from '../evaluate-cloudformation-template';
 
 export interface NestedStackTemplates {
