+++ packages/aws-cdk/lib/api/logs/logs-monitor.ts
@@ -2,7 +2,7 @@ import * as util from 'util';
 import * as cxapi from '@aws-cdk/cx-api';
 import * as chalk from 'chalk';
 import { info, error } from '../../logging';
-import { flatten } from '../../util/arrays';
+import { flatten } from '../../util';
 import type { SDK } from '../aws-auth';
 
 /**
