+++ packages/aws-cdk/lib/api/util/placeholders.ts
@@ -1,5 +1,5 @@
 import { type Environment, EnvironmentPlaceholders } from '@aws-cdk/cx-api';
-import { Branded } from '../../util/type-brands';
+import type { Branded } from '../../util';
 import type { SdkProvider } from '../aws-auth/sdk-provider';
 import { Mode } from '../plugin/mode';
