+++ packages/aws-cdk/lib/util.ts
@@ -1,0 +1,1 @@
+export * from '../../@aws-cdk/tmp-toolkit-helpers/src/util';
