+++ packages/aws-cdk/test/api/aws-auth/fake-sts.ts
@@ -2,7 +2,7 @@ import { Tag } from '@aws-sdk/client-sts';
 import * as nock from 'nock';
 import * as uuid from 'uuid';
 import * as xmlJs from 'xml-js';
-import { formatErrorMessage } from '../../../lib/util/format-error';
+import { formatErrorMessage } from '../../../lib/util';
 
 interface RegisteredIdentity {
   readonly account: string;
